fix: access control and layout
